src/backend/requirements.txt
fastapi==0.128.4
uvicorn[standard]==0.40.0
sqlalchemy==2.0.46
mysql-connector-python==9.5.0
pydantic==2.12.5
python-dotenv==1.2.1
python-multipart==0.0.20
reportlab==4.4.4
python-docx==1.2.0
